% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBuilder.R
\name{add_class_metadata}
\alias{add_class_metadata}
\title{Adding class metadata to vertices}
\usage{
add_class_metadata(
   g, 
   mets
)
}
\arguments{
\item{g}{iGraph object obtained from to_graph() or get_graph()}

\item{mets}{iGraph identifiers of metabolites}
}
\description{
Adding class metadata to vertices
}
