% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBuilder.R
\name{add_closeness}
\alias{add_closeness}
\title{Add closeness to your graph}
\usage{
add_closeness(
   graph
)
}
\arguments{
\item{graph}{igraph object representing your graph}
}
\description{
Add closeness to your graph
}
