% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBuilder.R
\name{add_communities}
\alias{add_communities}
\title{Identify communities in the graph}
\usage{
add_communities(
   graph
)
}
\arguments{
\item{graph}{igraph object representing your graph}
}
\description{
Identify communities in the graph
}
