% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBuilder.R
\name{add_gos}
\alias{add_gos}
\title{Assign Gene Ontologies to vertices}
\usage{
add_gos(
   graph,
   protein_go_df,
   verbose = T
)
}
\arguments{
\item{graph}{igraph object representing your graph}

\item{protein_go_df}{placeholder}

\item{verbose}{Present a warning when no proteins are found?}
}
\description{
Assign Gene Ontologies to vertices
}
\examples{
\dontrun{
graph <- example_graph()
graph <- add_gos(graph)
}
}
