% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBuilder.R
\name{add_metadata}
\alias{add_metadata}
\title{Adding metadata to igraph object}
\usage{
add_metadata(graph)
}
\arguments{
\item{graph}{iGraph object obtained from to_graph() or get_graph()}
}
\description{
Adding metadata to igraph object
}
