% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBuilder.R
\name{add_metadata_to_graph}
\alias{add_metadata_to_graph}
\title{Adding metadata to igraph object}
\usage{
add_metadata_to_graph(
   g
)
}
\arguments{
\item{g}{iGraph object obtained from to_graph() or get_graph()}
}
\description{
Adding metadata to igraph object
}
