% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBuilder.R
\name{add_node_pvalues}
\alias{add_node_pvalues}
\title{calculate all node-based pvalues}
\usage{
add_node_pvalues(
   graph,
   protein_go_df,
   order = 1
)
}
\arguments{
\item{graph}{placeholder}

\item{order}{placeholder}

\item{protein_go_df}{placeholder}
}
\description{
calculate all node-based pvalues
}
