% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBuilder.R
\name{add_node_types}
\alias{add_node_types}
\title{Add the type of each node}
\usage{
add_node_types(graph)
}
\arguments{
\item{graph}{iGraph object obtained from to_graph() or get_graph()}
}
\description{
Add the type of each node
}
