% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBuilder.R
\name{add_precision}
\alias{add_precision}
\title{add precision score to node}
\usage{
add_precision(graph, omit_lipids = F)
}
\arguments{
\item{graph}{iGraph object obtained from to_graph() or get_graph()}
}
\description{
add precision score to node
}
