% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBuilder.R
\name{add_superclass_metadata}
\alias{add_superclass_metadata}
\title{Adding superclass metadata to vertices}
\usage{
add_superclass_metadata(
   g,
   mets
)
}
\arguments{
\item{g}{iGraph object obtained from to_graph() or get_graph()}

\item{mets}{iGraph identifiers of metabolites}
}
\description{
Adding superclass metadata to vertices
}
