% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBuilder.R
\name{add_vertice_colors}
\alias{add_vertice_colors}
\title{Adding colors to vertices of the graph}
\usage{
add_vertice_colors(
   graph,
)
}
\arguments{
\item{graph}{iGraph object obtained from to_graph() or get_graph()}
}
\description{
Adding colors to vertices of the graph
}
