% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{adjust_folder}
\alias{adjust_folder}
\title{adjust folder path in given options}
\usage{
adjust_folder(
   options
)
}
\arguments{
\item{options}{List object constructed from a YAML file}
}
\description{
adjust folder path in given options
}
