% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataController.R
\name{build_button}
\alias{build_button}
\title{Build a new graph}
\usage{
build_button(
   filter
)
}
\arguments{
\item{filter}{String containing the filter to be used / searched.}
}
\description{
Build a new graph
}
