% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrappers.R
\name{calculate_all_node_pvalues}
\alias{calculate_all_node_pvalues}
\title{Run calculate all metabolite-go pvalues}
\usage{
calculate_all_node_pvalues(config_path, order = 1)
}
\description{
Run calculate all metabolite-go pvalues
}
