% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBuilder.R
\name{calculate_pvalues}
\alias{calculate_pvalues}
\title{Calculate all GO pvalues in the graph}
\usage{
calculate_pvalues(
   g
)
}
\arguments{
\item{g}{iGraph object obtained from to_graph() or get_graph()}
}
\description{
Calculate all GO pvalues in the graph
}
