% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{change_index}
\alias{change_index}
\title{Change rownames (index) of a dataframe}
\usage{
change_index(
   df,
   index_column
)
}
\arguments{
\item{df}{Dataframe of which the rownames need to be set}

\item{index_column}{String of ccolumn name which needs to be set as rownames}
}
\description{
Change rownames (index) of a dataframe
}
