% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBuilder.R
\name{construct_metabolite_go_network}
\alias{construct_metabolite_go_network}
\title{Construct a network of metabolites and GOs}
\usage{
construct_metabolite_go_network(
   g,
   mets
)
}
\arguments{
\item{g}{iGraph object obtained from to_graph() or get_graph()}

\item{mets}{iGraph identifiers of metabolites}
}
\description{
Construct a network of metabolites and GOs
}
