% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{convert_ids_to_names}
\alias{convert_ids_to_names}
\title{Convert identifiers to names}
\usage{
convert_ids_to_names(
   ids
)
}
\arguments{
\item{ids}{Vector of protein / metabolite identifiers}
}
\description{
Convert identifiers to names
}
