% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{convert_names_to_ids}
\alias{convert_names_to_ids}
\title{Convert names to identifiers}
\usage{
convert_names_to_ids(
   names
)
}
\arguments{
\item{names}{Vector of protein / metabolite names}
}
\description{
Convert names to identifiers
}
