% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrappers.R
\name{example_graph}
\alias{example_graph}
\title{Produce the example graph}
\usage{
example_graph(omit_lipids = F)
}
\arguments{
\item{omit_lipids}{Boolean value}
}
\description{
Produce the example graph
}
\examples{
\dontrun{
graph <- example_graph()
plot(graph)
}
}
