% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{filter_on_confidence}
\alias{filter_on_confidence}
\title{Filter by interaction confidence}
\usage{
filter_on_confidence(
   df,
   conf
)
}
\arguments{
\item{df}{Dataframe containing protein-protein interactions found}

\item{conf}{Integer between 0 (low) and 1000 (high) determining the confidence level}
}
\description{
Filter protein-protein interactions by the
interaction confidence as determined by StringDB (1-1000)
}
