% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBuilder.R
\name{fishers_test}
\alias{fishers_test}
\title{Calculate GO pvalues using Fisher's test}
\usage{
fishers_test(
   id,
   all_go_in_network,
   all_go
)
}
\arguments{
\item{id}{String Gene Ontology identifier}

\item{all_go_in_network}{Table of Gene Ontology counts in the current network}

\item{all_go}{Table of Gene Ontology counts in the database}
}
\description{
Calculate GO pvalues using Fisher's test
}
