% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBuilder.R
\name{get_2d_scatter}
\alias{get_2d_scatter}
\title{Produce 2D pvalue-closeness scatter plot}
\usage{
get_2d_scatter(
   g
)
}
\arguments{
\item{g}{iGraph object obtained from to_graph() or get_graph()}
}
\description{
Produce 2D pvalue-closeness scatter plot
}
