% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{get_all_interactions}
\alias{get_all_interactions}
\title{Get all type of interactions}
\usage{
get_all_interactions(
   ids,
   mode = "both",
   conf = 0
)
}
\arguments{
\item{ids}{Vector of identifiers that may contain metabolites and/or proteins}

\item{mode}{String determining the mode, see examples}

\item{conf}{Integer between 0 (low) and 1000 (high) determining the confidence level}
}
\description{
Get all type of interactions between given identifiers.
}
