% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataController.R
\name{get_barplot}
\alias{get_barplot}
\title{Get barplot for Closeness scores}
\usage{
get_barplot(
   g
)
}
\arguments{
\item{g}{iGraph object obtained from to_graph() or get_graph()}
}
\description{
Get barplot for Closeness scores
}
