% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{get_cellular_location}
\alias{get_cellular_location}
\title{Get the cellular location of metabolites}
\usage{
get_cellular_location(
   ids
)
}
\arguments{
\item{ids}{Vector of metabolite identifiers}
}
\description{
Get the cellular location of metabolites
}
