% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBuilder.R
\name{get_cofactor_edge_ids}
\alias{get_cofactor_edge_ids}
\title{Get indexes of cofactor edges by identifier}
\usage{
get_cofactor_edge_ids(
   g,
   cofactors
)
}
\arguments{
\item{g}{iGraph object obtained from to_graph() or get_graph()}

\item{cofactors}{Vector of metabolite cofactors}
}
\description{
Get indexes of cofactor edges by identifier
}
