% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{get_cofactors}
\alias{get_cofactors}
\title{Get metabolite cofactors for given protein identifiers}
\usage{
get_cofactors(
   ids
)
}
\arguments{
\item{ids}{Vector of protein identifiers}
}
\description{
Get metabolite cofactors for given protein identifiers
}
