% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBuilder.R
\name{get_edge_ids}
\alias{get_edge_ids}
\title{Get edge ids between nodes of interest using shortest paths.}
\usage{
get_edge_ids(
   g,
   combinations
)
}
\arguments{
\item{g}{iGraph object obtained from to_graph() or get_graph()}
}
\description{
Get edge ids between nodes of interest using shortest paths.
}
