% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBuilder.R
\name{get_edge_shapes}
\alias{get_edge_shapes}
\title{Return the coordinates for edges}
\usage{
get_edge_shapes(
   g,
   coord
)
}
\arguments{
\item{g}{iGraph object obtained from to_graph() or get_graph()}

\item{coord}{Dataframe containing coordinates of the nodes}
}
\description{
Return the coordinates for edges
}
