% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataController.R
\name{get_edge_table}
\alias{get_edge_table}
\title{Get datatable for edges}
\usage{
get_edge_table(
   g
)
}
\arguments{
\item{g}{iGraph object obtained from to_graph() or get_graph()}
}
\description{
Get datatable for edges
}
