% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBuilder.R
\name{get_enzyme_vertice_ids}
\alias{get_enzyme_vertice_ids}
\title{Get indexes of enzymes vertices by identifier}
\usage{
get_enzyme_vertice_ids(
   g
)
}
\arguments{
\item{g}{iGraph object obtained from to_graph() or get_graph()}
}
\description{
Get indexes of enzymes vertices by identifier
}
