% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{get_enzymes}
\alias{get_enzymes}
\title{Get enzyme EC numbers for given protein identifiers}
\usage{
get_enzymes(
   ids
)
}
\arguments{
\item{ids}{Vector of protein identifiers}
}
\description{
Get enzyme EC numbers for given protein identifiers
}
