% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataController.R
\name{get_go_barplot}
\alias{get_go_barplot}
\title{Get Plotly barplot for Gene Ontologies}
\usage{
get_go_barplot(
    g
)
}
\arguments{
\item{g}{iGraph object obtained from to_graph() or get_graph()}
}
\description{
Get Plotly barplot for Gene Ontologies
}
