% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{get_go_ids_by_go}
\alias{get_go_ids_by_go}
\title{Get Gene Ontology identifiers by name}
\usage{
get_go_ids_by_go(
   gos
)
}
\arguments{
\item{gos}{Vector of Gene Ontology names}
}
\description{
Get Gene Ontology identifiers by name
}
