% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{get_go_terms}
\alias{get_go_terms}
\title{Get GO terms}
\usage{
list_go_terms()
}
\description{
Returns all possible Gene Ontology search terms found in the data
}
\examples{
go_vec <- get_go_terms()
}
