% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{get_gos}
\alias{get_gos}
\title{Get Gene Ontologys}
\usage{
get_gos(
   options
)
}
\arguments{
\item{options}{YAML list containing field-value pairs.
See the ReadMe for mandatory fields. Returns vector of GO IDs.
options = list(GO_ID = "GO:0002376")
go_vec <- get_gos(options)}
}
\description{
Retrieves GO descendants of the given GO identifier.
}
