% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBuilder.R
\name{get_gos_per_cluster}
\alias{get_gos_per_cluster}
\title{Assign GOs to metabolites & proteins}
\usage{
get_gos_per_cluster(
   g
)
}
\arguments{
\item{g}{iGraph object obtained from to_graph() or get_graph()}
}
\description{
Assign GOs to metabolites & proteins
}
