% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrappers.R
\name{get_graph}
\alias{get_graph}
\title{Get igraph object}
\usage{
get_graph(
   filter,
   neighbours = 0,
   max_neighbours = Inf,
   type = "Gene Ontology",
   simple = F,
   search_mode = "Interacts"
)
}
\arguments{
\item{filter}{String containing the search term.}

\item{neighbours}{Integer containing the number of neighbours to be found}

\item{max_neighbours}{Integer representing the maximum number of edges for each neighbour}

\item{simple}{Boolean value indicating to return a barebone graph or including metadata.}

\item{type}{String containing the type of search.}

\item{search_mode}{String of either 'Interacts' or 'Between'. Interacts finds the first neighbour
of the given search, while Between only returns interactions between the proteins / metabolites given.}
}
\description{
Get igraph object
}
\examples{
# Construct a GO graph
g <- get_graph("microglial cell activation")
g <- get_graph("microglial cell activation", 0, "Gene Ontology")

# Construct a graph using Metabolites and/or Proteins
g <- get_graph("L-Asparagine", type = "Metabolites/Proteins")
}
