% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrappers.R
\name{get_graph_summary}
\alias{get_graph_summary}
\title{Get a summary of the graph}
\usage{
get_graph_summary(
   graph
)
}
\arguments{
\item{graph}{iGraph object obtained from to_graph() or get_graph()}
}
\description{
Get a summary of the graph
}
