% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataController.R
\name{get_heatmap_plot}
\alias{get_heatmap_plot}
\title{Get Plotly heatmap of distances}
\usage{
get_heatmap_plot(
   g
)
}
\arguments{
\item{g}{iGraph object obtained from to_graph() or get_graph()}
}
\description{
Get Plotly heatmap of distances
}
