% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{get_ids_from_class}
\alias{get_ids_from_class}
\title{Get metabolite identifiers by class}
\usage{
get_ids_from_class(
   filter
)
}
\arguments{
\item{filter}{Vector of metabolite class names}
}
\description{
Get metabolite identifiers by class
}
