% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{get_ids_from_pathways}
\alias{get_ids_from_pathways}
\title{Get metabolite identifiers by pathways}
\usage{
get_ids_from_pathways(
   filter
)
}
\arguments{
\item{filter}{Vector of pathway names}
}
\description{
Get metabolite identifiers by pathways
}
