% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{get_ids_from_superclass}
\alias{get_ids_from_superclass}
\title{Get metabolite identifiers by superclass}
\usage{
get_ids_from_superclass(
   filter
)
}
\arguments{
\item{filter}{Vector of metabolite superclass names}
}
\description{
Get metabolite identifiers by superclass
}
