% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{get_interactions}
\alias{get_interactions}
\title{Get interactions}
\usage{
get_interactions(
   df,
   ids,
   mode
)
}
\arguments{
\item{df}{Dataframe of interactions containing a 'From' and 'To' column}

\item{ids}{Vector of identifiers to be searched}

\item{mode}{String determining the mode, see examples}
}
\description{
Get interactions
}
