% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrappers.R
\name{get_metabolite_go_pvalues}
\alias{get_metabolite_go_pvalues}
\title{Return a long-format DataFrame with GO-Pvalue combinations per metabolite}
\usage{
get_metabolite_go_pvalues(
   graph
)
}
\arguments{
\item{graph}{igraph object}
}
\description{
Return a long-format DataFrame with GO-Pvalue combinations per metabolite
}
