% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{get_metabolite_ids}
\alias{get_metabolite_ids}
\title{Get metabolite identifiers}
\usage{
get_metabolite_ids(
   names
)
}
\arguments{
\item{names}{Vector of metabolite names}
}
\description{
Get metabolite identifiers
}
