% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrappers.R
\name{get_metabolite_metadata}
\alias{get_metabolite_metadata}
\title{Get metadata of metabolites}
\usage{
get_metabolite_metadata(
   graph,
   metadata
)
}
\arguments{
\item{graph}{igraph object obtained from 'get_graph()'}

\item{metadata}{String containing column name of metadata to obtain}
}
\description{
Get metadata of metabolites
}
\examples{
# Get GOs associated with metabolites
g <- get_graph("microglial cell activation")
get_metabolite_metadata(g, "go")
}
