% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{get_metabolite_names}
\alias{get_metabolite_names}
\title{Get metabolite names}
\usage{
get_metabolite_names(
   metabolite_ids
)
}
\arguments{
\item{metabolite_ids}{Vector of metabolite IDs}
}
\description{
Get metabolite names
}
