% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{get_mm_interactions}
\alias{get_mm_interactions}
\title{Get metabolite-metabolite interactions}
\usage{
get_mm_interactions(
   ids,
   mode = "both"
)
}
\arguments{
\item{ids}{Vector of identifiers that may contain metabolites and/or proteins}

\item{mode}{String determining the mode, see examples}
}
\description{
Get metabolite-metabolite interactions
}
