% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{get_n_neighbours}
\alias{get_n_neighbours}
\title{Get N neirest neighbours with a maximum number of edges}
\usage{
get_n_neighbours(
   df,
   n,
   max
)
}
\arguments{
\item{df}{Dataframe of current interactions}

\item{n}{Integer of number of neighbours}

\item{max}{Integer of maximum number of edges per node.}
}
\description{
Get N neirest neighbours with a maximum number of edges
}
