% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrappers.R
\name{get_node_summary}
\alias{get_node_summary}
\title{Get a summary of all nodes in a graph}
\usage{
get_node_summary(
   graph
)
}
\arguments{
\item{graph}{iGraph object obtained from to_graph() or get_graph()}
}
\description{
Get a summary of all nodes in a graph
}
