% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataController.R
\name{get_node_table}
\alias{get_node_table}
\title{Get datatable for nodes}
\usage{
get_node_table(
   g
)
}
\arguments{
\item{g}{iGraph object obtained from to_graph() or get_graph()}
}
\description{
Get datatable for nodes
}
