% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{get_pm_interaction_ids}
\alias{get_pm_interaction_ids}
\title{Get protein-metabolite interactions}
\usage{
get_pm_interactions(
   ids,
   mode = "both"
)
}
\arguments{
\item{ids}{Vector of identifiers that may contain metabolites and/or proteins}

\item{mode}{String determining the mode, see examples}
}
\description{
Get protein-metabolite interactions
}
