% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBuilder.R
\name{get_pp_confidences}
\alias{get_pp_confidences}
\title{Get protein-protein interaction confidences}
\usage{
get_pp_confidences(
   g
)
}
\arguments{
\item{g}{iGraph object obtained from to_graph() or get_graph()}
}
\description{
Get protein-protein interaction confidences
}
