% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{get_pp_interactions}
\alias{get_pp_interactions}
\title{Get protein-protein interactions}
\usage{
get_pp_interactions(
   ids,
   mode = "both"
)
}
\arguments{
\item{ids}{Vector of identifiers that may contain metabolites and/or proteins}

\item{mode}{String determining the mode, see examples}
}
\description{
Get protein-protein interactions
}
