% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrappers.R
\name{get_process_summary}
\alias{get_process_summary}
\title{Get a summary of a process graph}
\usage{
get_process_summary(graph)
}
\arguments{
\item{graph}{iGraph object obtained from to_graph() or get_graph()}
}
\description{
Get a summary of a process graph
}
