% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{get_protein_ids}
\alias{get_protein_ids}
\title{Retrieve protein identifiers using names}
\usage{
get_protein_ids(names)
}
\arguments{
\item{names}{Vector of protein names}
}
\description{
Retrieve protein identifiers using names
}
