% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{get_protein_names}
\alias{get_protein_names}
\title{Get protein names}
\usage{
get_protein_names(
   protein_ids,
   synonym = False
)
}
\arguments{
\item{protein_ids}{Vector of protein IDs}

\item{synonym}{Boolean value, do synonyms need to be retrieved?}
}
\description{
Get protein names
}
