% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBuilder.R
\name{get_protein_vertice_ids}
\alias{get_protein_vertice_ids}
\title{Get indexes of proteins vertices by identifier}
\usage{
get_protein_vertice_ids(
   g
)
}
\arguments{
\item{g}{iGraph object obtained from to_graph() or get_graph()}
}
\description{
Get indexes of proteins vertices by identifier
}
