% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{get_proteins_by_go}
\alias{get_proteins_by_go}
\title{Return all proteins associated with a GO}
\usage{
get_proteins_by_go(names)
}
\arguments{
\item{names}{Vector of Gene Ontology names}
}
\description{
Return all proteins associated with a GO
}
