% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{get_proteins_by_goid}
\alias{get_proteins_by_goid}
\title{Get protein names by Gene Ontology identifiers}
\usage{
get_proteins_by_goid(
   ids
)
}
\arguments{
\item{ids}{Vector of Gene Ontology IDs}
}
\description{
Get protein names by Gene Ontology identifiers
}
