% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBuilder.R
\name{get_shortest_path_graph}
\alias{get_shortest_path_graph}
\title{Obtain shortest graph between nodes}
\usage{
get_shortest_path_graph(
   g,
   filter
)
}
\arguments{
\item{g}{iGraph object obtained from to_graph() or get_graph()}

\item{filter}{Names of protein/metabolites to search}
}
\description{
Obtain shortest graph between nodes
}
