% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{get_sorted_interaction_names}
\alias{get_sorted_interaction_names}
\title{Get names of all available proteins / metabolites}
\usage{
get_sorted_interaction_names()
}
\description{
Get names of all available proteins / metabolites
}
