% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{get_superclass}
\alias{get_superclass}
\title{Get superclasses of metabolites}
\usage{
get_superclass(
   ids
)
}
\arguments{
\item{ids}{Vector of metabolite identifiers}
}
\description{
Get superclasses of metabolites
}
