% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBuilder.R
\name{get_vertice_id}
\alias{get_vertice_id}
\title{Get index of vertice in graph by identifier}
\usage{
get_vertice_id(
   g,
   id
)
}
\arguments{
\item{g}{iGraph object obtained from to_graph() or get_graph()}

\item{id}{String identifier of metabolite or protein}
}
\description{
Get index of vertice in graph by identifier
}
