% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBuilder.R
\name{harmonized_closeness}
\alias{harmonized_closeness}
\title{Calculate the harmonized closeness of a graph}
\usage{
harmonized_closeness(
   g
)
}
\arguments{
\item{g}{iGraph object obtained from to_graph() or get_graph()}
}
\description{
Calculate the harmonized closeness of a graph
}
