% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI.R
\name{header_row}
\alias{header_row}
\title{Header row}
\usage{
header_row(title, subtitle)
}
\arguments{
\item{title}{Title of the page}

\item{subtitle}{subtitle of the page}
}
\value{
A fluidRow meant for each header of the page.
}
\description{
This function creates a header with a title and subtitle for each page
}
