% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBuilder.R
\name{inheritence_by_threshold}
\alias{inheritence_by_threshold}
\title{inheritance of gos}
\usage{
inheritence_by_threshold(graph, pvalues, threshold = 1, neighbours = 0)
}
\description{
inheritance of gos
}
