% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{load_data}
\alias{load_data}
\title{Load Data}
\usage{
load_data()
}
\description{
Variables and files are loaded to the global environment to
be used by other functions inside this package.
}
\examples{
load_data()
}
