% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataController.R
\name{load_interaction_data}
\alias{load_interaction_data}
\title{Load interaction data}
\usage{
load_interaction_data(
   options,
   prot_file = "Protein-protein.csv",
   full_load = TRUE
)
}
\arguments{
\item{options}{YAML list containing field-value pairs.}
}
\description{
Load interaction data
}
