% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBuilder.R
\name{metabolite_go_graph}
\alias{metabolite_go_graph}
\title{Construct a network of metabolites and GOs}
\usage{
metabolite_go_graph(
   graph
)
}
\arguments{
\item{graph}{iGraph object obtained from get_graph()}
}
\description{
Construct a network of metabolites and GOs
}
