% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{normalized}
\alias{normalized}
\title{Normalize vectors}
\usage{
normalized(
   vec,
   min = 0,
   max = 1
)
}
\arguments{
\item{vec}{Vector of numerical values}

\item{min}{Lower boundary after normalization}

\item{max}{Upper boundary after normalization}
}
\description{
Normalize vectors
}
