% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataController.R
\name{observe_inputs}
\alias{observe_inputs}
\title{Observe reactive inputs from the Shiny app}
\usage{
observe_inputs()
}
\description{
In the server, call observe_inputs() to synchronize inputs
in the Shiny app.
}
\examples{
observe_inputs()
}
