% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{read_file}
\alias{read_file}
\title{Read a file's contents}
\usage{
read_file(
  source,
  index_column
)
}
\arguments{
\item{source}{String path of the file location.}

\item{index_column}{(optional) String name of column
file <- read_file("metabolites.csv", "Name")}
}
\description{
Read a file's contents
}
