% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrappers.R
\name{run_preprocessing}
\alias{run_preprocessing}
\title{Run Preprocessing}
\usage{
run_preprocessing(
   config_path = "config.yaml"
)
}
\arguments{
\item{config_path}{placeholder}
}
\description{
Run Preprocessing
}
