% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{run_shiny}
\alias{run_shiny}
\title{Run Shiny App}
\usage{
run_shiny(
   browser = TRUE,
   port = 2222
)
}
\arguments{
\item{browser}{Should the browser be started?}

\item{port}{Which port to be used}
}
\description{
Run Shiny App
}
\examples{
# Start the Atlas in your browser
\dontrun{
run_shiny()
}
}
