% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Server.R
\name{server}
\alias{server}
\title{Shiny server}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{Reactive variable containing all inputs from the user}

\item{output}{Reactive variable containing all outputs to the server}

\item{session}{Reactive variable containing info about the specific session of the user}
}
\description{
Shiny server
}
