% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrappers.R
\name{textmining_analysis}
\alias{textmining_analysis}
\title{Perform text mining analysis}
\usage{
textmining_analysis(
   config_path = "config.yaml",
   font = 18
)
}
\arguments{
\item{config_path}{placeholder}

\item{font}{placeholder}
}
\description{
Perform text mining analysis
}
