% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBuilder.R
\name{to_graph}
\alias{to_graph}
\title{Convert interaction dataframe to iGraph object}
\usage{
to_graph(
   df
)
}
\arguments{
\item{df}{Dataframe of interactions between nodes}
}
\description{
Convert interaction dataframe to iGraph object
}
