% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBuilder.R
\name{to_plotly}
\alias{to_plotly}
\title{Convert graph object to an interactive Plotly}
\usage{
igraph_to_plotly2(
   g
)
}
\arguments{
\item{g}{iGraph object obtained from to_graph() or get_graph()}
}
\description{
Convert graph object to an interactive Plotly
}
