% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UI.R
\name{ui.top_row}
\alias{ui.top_row}
\title{Create the UI search row}
\usage{
ui.top_row(suffix)
}
\arguments{
\item{suffix}{The suffix is used to create unique objects. This suffix is based on the page they're
located. The main page has an empty suffix, while the rest has "-Heatmap" as suffix for example}
}
\value{
A fluidRow made for placement below the header_row in the visualization pages.
}
\description{
This function is used by the UI only. It creates the top fluidrow in the
visualization pages Network, Heatmap, Statistics & Data Table.
In this fluidrow 2 select inputs and 1 action button are located. These are shaped
by css and thus will look equal in each page for consistency
}
