% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBuilder.R
\name{add_gos}
\alias{add_gos}
\title{Assign Gene Ontologies to vertices}
\usage{
add_gos(
   graph,
   verbose = T
)
}
\arguments{
\item{graph}{igraph object representing your graph}

\item{verbose}{Present a warning when no proteins are found?}
}
\description{
Assign Gene Ontologies to vertices
}
