% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBuilder.R
\name{add_node_pvalues}
\alias{add_node_pvalues}
\title{calculate all node-based pvalues}
\usage{
add_node_pvalues(graph, order = 1)
}
\arguments{
\item{graph}{iGraph object obtained from to_graph() or get_graph()}

\item{order}{placeholder}
}
\description{
calculate all node-based pvalues
}
