% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataController.R
\name{data_filter}
\alias{data_filter}
\title{Retrieve data given a search}
\usage{
data_filter(
   filter,
   neighbours = 0,
   max_neighbours = 0,
   type = "Gene Ontology",
   search_mode = "Interacts"
)
}
\description{
Retrieve data given a search
}
