% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper_functions.R
\name{get_graph}
\alias{get_graph}
\title{Get igraph object}
\usage{
get_graph(
   filter,
   neighbours = 0,
   max_neighbours = Inf,
   type = "Gene Ontology",
   simple = F
)
}
\arguments{
\item{filter}{String containing the search term.}

\item{neighbours}{Integer containing the number of neighbours to be found}

\item{type}{String containing the type of search.}
}
\description{
Get igraph object
}
\examples{
# Construct a GO graph
g <- get_graph("microglial cell activation")
g <- get_graph("microglial cell activation", 0, "Gene Ontology")

# Construct a graph using Metabolites and/or Proteins
g <- get_graph("L-Asparagine", type = "Metabolites/Proteins")
}
