% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{get_protein_ids}
\alias{get_protein_ids}
\title{Get protein identifiers}
\usage{
get_protein_ids(
   protein_names
)
}
\arguments{
\item{protein_names}{Vector of protein names}
}
\description{
Get protein identifiers
}
