% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{get_protein_names}
\alias{get_protein_names}
\title{Retrieve protein names using identifiers}
\usage{
get_protein_names(
   protein_ids
)
}
\arguments{
\item{protein_ids}{Vector of protein IDs}
}
\description{
Retrieve protein names using identifiers
}
