% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrappers.R
\name{load_data}
\alias{load_data}
\title{Load Data}
\usage{
load_data(
   config,
   neighbours = 0,
   confidence = 0,
   full = TRUE,
   verbose = TRUE
)
}
\arguments{
\item{config}{String path to a YAML configuration file}

\item{neighbours}{Integer value representing the number of 'steps' allowed for protein-protein interactions}

\item{confidence}{placeholder}

\item{full}{placeholder}

\item{verbose}{placeholder}
}
\description{
Variables and files are loaded to the global environment to
be used by other functions inside this package.
}
