% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataController.R
\name{load_interaction_data}
\alias{load_interaction_data}
\title{Load interaction data}
\usage{
load_interaction_data(
   options,
   prot_file = "Protein-protein.csv",
   full_load = TRUE
)
}
\arguments{
\item{options}{YAML list containing field-value pairs.}

\item{prot_file}{String name of the protein file}

\item{full_load}{Boolean value for loading everything or just the protein-protein interaction file}
}
\description{
Load interaction data
}
