% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrappers.R
\name{metabolic_process_summary}
\alias{metabolic_process_summary}
\title{Return a long-format DataFrame with GO-Pvalue combinations per metabolite}
\usage{
metabolic_process_summary(
   graph,
   p = 1,
   centrality = 0
)
}
\arguments{
\item{graph}{igraph object}

\item{p}{placeholder}

\item{centrality}{placeholder}
}
\description{
Return a long-format DataFrame with GO-Pvalue combinations per metabolite
}
