% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBuilder.R
\name{remove_unconnected}
\alias{remove_unconnected}
\title{Remove unconnected nodes}
\usage{
remove_unconnected(
   graph
)
}
\arguments{
\item{graph}{igraph object representing your graph}
}
\description{
Remove unconnected nodes
}
