% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphBuilder.R
\name{to_plotly}
\alias{to_plotly}
\title{Convert igraph object to an interactive Plotly}
\usage{
to_plotly(
   graph
)
}
\arguments{
\item{graph}{iGraph object obtained from get_graph()}
}
\description{
Convert igraph object to an interactive Plotly
}
